+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+- `Moar.URI.valid?` will now return `true` when given a `mailto:` URI, as long as its path is not blank.
+- **Breaking changes:** in previous versions, `Moar.URI.valid?` would return `false` when given a `mailto:` URI. To ensure
+  that a URI is a valid *web* URI, use `Moar.URI.web_url?`
+
 ## 2.7.0
 
 - Add `:to_string` shortcut for `assert_eq` which converts both terms to strings before comparing.